pyproject.toml
[project]
name = "enterprise-database-migration"
version = "0.1.0"
description = "PostgreSQL Enterprise Database Migration System - Efficient and reliable database migration tool"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Yves Marinho", email = "yves.marinho@example.com"}
]
keywords = ["postgresql", "database", "migration", "enterprise", "data-transfer"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

# Dependências essenciais (baseadas na análise do código)
dependencies = [
    "psycopg2-binary>=2.9.7",      # PostgreSQL connector (usado extensivamente)
    "SQLAlchemy>=2.0.23",          # ORM (usado em cleanup e migrations)
    "colorama>=0.4.6",             # Terminal colors (orchestrator_pure_python)
    "mysql-connector-python>=8.1.0", # MySQL connector (testes de validação)
    "jsonschema>=4.20.0",          # JSON validation
]

[project.optional-dependencies]
# Dependências de desenvolvimento (opcional)
dev = [
    "pytest>=8.0.0",          # Versão mais moderna com Python 3.11+
    "pytest-asyncio>=0.23.0", # Melhor suporte async com Python moderno
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",          # Versão mais recente com melhor performance
    "isort>=5.13.2",
    "flake8>=7.0.0",          # Versão mais moderna
    "pylint>=3.1.0",          # Melhor suporte Python 3.11+
    "mypy>=1.9.0",            # Versão mais recente com melhor type checking
    "pre-commit>=3.8.0",      # Versão mais estável
]

# Dependências para funcionalidades avançadas (futuras)
advanced = [
    "rich>=13.7.0",           # Enhanced CLI formatting
    "click>=8.1.7",           # CLI framework
    "structlog>=23.2.0",      # Structured logging
    "python-dotenv>=1.0.0",   # Environment variables
    "PyYAML>=6.0.1",          # YAML configuration
    "pydantic>=2.5.2",        # Data validation
]

# Dependências para monitoramento (futuras)
monitoring = [
    "memory-profiler>=0.61.0",
    "psutil>=5.9.6",
    "prometheus-client>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/yvesmarinho/enterprise-database-migration"
Repository = "https://github.com/yvesmarinho/enterprise-database-migration"
Issues = "https://github.com/yvesmarinho/enterprise-database-migration/issues"
Documentation = "https://github.com/yvesmarinho/enterprise-database-migration/blob/main/README.md"

[project.scripts]
# Scripts de linha de comando
migrate-db = "run_migration:main"
db-orchestrator = "orchestrator_pure_python:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Especifica quais arquivos/diretórios incluir no build
packages = [
    "core",
    "utils",
    "cleanup",
    "validation",
]
# Inclui arquivos Python na raiz do projeto e configurações
include = [
    "*.py",
    "config/*",
    "secrets/*.json",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.coverage.run]
source = ["core", "utils", "cleanup", "validation"]
omit = [
    "*/test*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
]
